Day 2 – Go: Improve Solution 2 by only checking neighbouring indexes.
As at this point, we know that the step between the previousLevel and the currentLevel has caused an invalid step, we can try removing the index before the last, and attempting to remove the current index.

If you imagine an array [0,1,0,5], the step from `1` back to `0` would fail, the failed index would be `1`, we should first try removing the `0` index, that doesn't give us a solution, we should then try removing index `2`, which does solve the problem.

Again, with [2,1,2,5], if we remove index `0` now, we get a solution.

Finally, for the test indexes [0,1], we check the removal of both the starting numbers as an edge case, as the sequence has just begun and there's immediately a failure, it's possible that either of these could be breaking the pattern that is not yet established (as the first pass of the method would've returned before it could access the rest of the sequence).